# internal/keyspan: add small performance improvements