**internal/keyspan: gate invariant checks behind build tag**

Gate invariant checks within the InterleavingIter behind a build tag.

**internal/keyspan: use pointer receiver for most Span methods**

Calls into Span methods in hot paths (especially `Span.Valid()`) were prominent
in CPU profiles. Using pointer receivers reduces the amount of memcpy-ing, and
allows the compiler to inline.

----

```
name                                                            old time/op  new time/op  delta
IteratorScan/keys=100,r-amp=1,key-types=points-only-10          5.87µs ± 0%  5.85µs ± 0%   -0.24%  (p=0.007 n=8+9)
IteratorScan/keys=100,r-amp=1,key-types=points-and-ranges-10    8.81µs ± 3%  7.55µs ± 0%  -14.37%  (p=0.000 n=10+8)
IteratorScan/keys=100,r-amp=3,key-types=points-only-10          10.4µs ± 3%  10.4µs ± 0%     ~     (p=1.000 n=10+10)
IteratorScan/keys=100,r-amp=3,key-types=points-and-ranges-10    13.8µs ± 1%  12.3µs ± 0%  -10.70%  (p=0.000 n=9+10)
IteratorScan/keys=100,r-amp=7,key-types=points-only-10          15.7µs ± 2%  15.5µs ± 1%     ~     (p=0.060 n=10+10)
IteratorScan/keys=100,r-amp=7,key-types=points-and-ranges-10    19.0µs ± 2%  17.4µs ± 0%   -8.44%  (p=0.000 n=10+9)
IteratorScan/keys=100,r-amp=10,key-types=points-only-10         19.2µs ± 1%  18.9µs ± 0%   -1.30%  (p=0.000 n=8+10)
IteratorScan/keys=100,r-amp=10,key-types=points-and-ranges-10   22.9µs ± 2%  21.0µs ± 0%   -8.21%  (p=0.000 n=10+10)
IteratorScan/keys=1000,r-amp=1,key-types=points-only-10         44.8µs ± 1%  44.3µs ± 0%   -0.96%  (p=0.000 n=9+9)
IteratorScan/keys=1000,r-amp=1,key-types=points-and-ranges-10   71.5µs ± 3%  59.7µs ± 0%  -16.56%  (p=0.000 n=10+10)
IteratorScan/keys=1000,r-amp=3,key-types=points-only-10         77.4µs ± 0%  76.4µs ± 0%   -1.28%  (p=0.000 n=8+10)
IteratorScan/keys=1000,r-amp=3,key-types=points-and-ranges-10    104µs ± 1%    91µs ± 0%  -12.60%  (p=0.000 n=10+8)
IteratorScan/keys=1000,r-amp=7,key-types=points-only-10          104µs ± 1%   104µs ± 0%     ~     (p=0.579 n=10+10)
IteratorScan/keys=1000,r-amp=7,key-types=points-and-ranges-10    133µs ± 0%   119µs ± 1%  -10.50%  (p=0.000 n=10+10)
IteratorScan/keys=1000,r-amp=10,key-types=points-only-10         118µs ± 1%   117µs ± 1%   -0.75%  (p=0.004 n=10+9)
IteratorScan/keys=1000,r-amp=10,key-types=points-and-ranges-10   145µs ± 1%   130µs ± 0%  -10.35%  (p=0.000 n=10+8)
IteratorScan/keys=10000,r-amp=1,key-types=points-only-10         418µs ± 0%   422µs ± 2%     ~     (p=0.063 n=9+9)
IteratorScan/keys=10000,r-amp=1,key-types=points-and-ranges-10   667µs ± 2%   574µs ± 3%  -14.03%  (p=0.000 n=10+9)
IteratorScan/keys=10000,r-amp=3,key-types=points-only-10         719µs ± 1%   715µs ± 2%     ~     (p=0.661 n=9+10)
IteratorScan/keys=10000,r-amp=3,key-types=points-and-ranges-10   973µs ± 2%   846µs ± 1%  -13.12%  (p=0.002 n=10+4)
```